+++ src/core/beacon.rs
@@ -1,6 +1,12 @@
 use super::chain::ChainHandler;
 use super::multibeacon::BeaconHandler;
 
+use crate::dkg::actions_active::ActionsActive;
+use crate::dkg::actions_passive::ActionsPassive;
+use crate::dkg::process::SeenPackets;
+use crate::dkg::store::DkgStore;
+use crate::dkg::ActionsError;
+
 use crate::key::keys::Identity;
 use crate::key::keys::Pair;
 use crate::key::store::FileStore;
@@ -10,20 +16,24 @@ use crate::key::toml::Toml;
 use crate::key::PointSerDeError;
 use crate::key::Scheme;
 
-use crate::dkg::process::Actions;
-use crate::dkg::process::DkgProcess;
 use crate::store::memstore::MemStore;
 use crate::store::ChainStore;
 use crate::store::NewStore;
 
+use crate::protobuf::dkg::packet::Bundle;
+use crate::protobuf::dkg::DkgStatusResponse;
 use crate::protobuf::drand::ChainInfoPacket;
 use crate::protobuf::drand::IdentityResponse;
+
+use crate::net::utils::Callback;
+use crate::transport::dkg::Command;
+use crate::transport::dkg::GossipPacket;
+use crate::transport::dkg::Participant;
 
 use energon::drand::traits::BeaconDigest;
 use std::path::Path;
 use std::sync::Arc;
 use tokio::sync::mpsc;
-use tokio::sync::oneshot;
 use tokio_util::task::TaskTracker;
 use tracing::debug;
 use tracing::error;
@@ -32,9 +42,6 @@ use tracing::info_span;
 use tracing::Span;
 
 pub const DEFAULT_BEACON_ID: &str = "default";
-
-/// Callback for [`BeaconCmd`]
-pub type Callback<T, E> = oneshot::Sender<Result<T, E>>;
 
 /// There is a direct relationship between an empty string and the reserved id "default".
 pub fn is_default_beacon_id(beacon_id: &str) -> bool {
@@ -66,14 +73,22 @@ impl BeaconID {
 
 /// Commands for communication between server-side and [`BeaconProcess`]
 pub enum BeaconCmd {
-    Shutdown(Callback<(), ()>),
+    Shutdown(Callback<(), ShutdownError>),
     IdentityRequest(Callback<IdentityResponse, PointSerDeError>),
-    Sync(Callback<Arc<ChainStore>, &'static str>),
-    ChainInfo(Callback<ChainInfoPacket, &'static str>),
+    Sync(Callback<Arc<ChainStore>, SyncError>),
+    ChainInfo(Callback<ChainInfoPacket, ChainInfoError>),
     DkgActions(Actions),
 }
 
-/// BeaconProcess is the main logic of the BeaconID. It reads the keys / group file, it
+/// Subcommand for DKG actions [`BeaconCmd::DkgActions`]
+pub enum Actions {
+    Gossip(GossipPacket, Callback<(), ActionsError>),
+    Command(Command, Callback<(), ActionsError>),
+    Broadcast(Bundle, Callback<(), ActionsError>),
+    Status(Callback<DkgStatusResponse, ActionsError>),
+}
+
+/// BeaconProcess is the main logic of the BeaconID instance. It reads the keys / group file, it
 /// can start the DKG, read/write shares to files and can initiate/respond to tBLS
 /// signature requests.
 pub struct BeaconProcess<S: Scheme> {
@@ -81,73 +96,58 @@ pub struct BeaconProcess<S: Scheme> {
     tracker: TaskTracker,
     fs: FileStore,
     keypair: Pair<S>,
-    span: Span,
     chain_handler: ChainHandler,
     chain_store: Arc<MemStore>,
+    dkg_store: DkgStore,
+    log: Span,
 }
 
 impl<S: Scheme> BeaconProcess<S> {
     fn new(fs: FileStore, pair: PairToml, id: &str) -> Result<Self, FileStoreError> {
         let keypair: Pair<S> = Toml::toml_decode(&pair).ok_or(FileStoreError::TomlError)?;
-        let span = info_span!(
-            "",
-            id = format!("{}.{id}", keypair.public_identity().address())
-        );
 
         let chain_handler = ChainHandler::try_init(&fs, id)?;
         let chain_store = Arc::new(ChainStore::new(&fs.db_path(), S::Beacon::is_chained())?);
+        let is_fresh = fs.is_fresh_run()?;
+        let dkg_store = DkgStore::init::<S>(fs.beacon_path.as_path(), is_fresh, id)?;
+        let log = info_span!(
+            "",
+            id = format!("{}.{id}", keypair.public_identity().address())
+        );
+        if is_fresh {
+            info!(parent: &log, "beacon id [{id}]: will run as fresh install -> expect to run DKG.");
+        }
         let process = Self {
             beacon_id: BeaconID::new(id),
             fs,
             chain_store,
             chain_handler,
             keypair,
-            span,
             tracker: TaskTracker::new(),
+            dkg_store,
+            log,
         };
 
         Ok(process)
     }
 
     pub fn run(fs: FileStore, pair: PairToml, id: &str) -> Result<BeaconHandler, FileStoreError> {
-        let is_new_node = fs.is_fresh_run()?;
-        let bp = Box::new(Self::new(fs, pair, id)?);
-        let dkg = DkgProcess::init(&bp, is_new_node)?;
-
+        let bp = Self::new(fs, pair, id)?;
         let tracker = bp.tracker().to_owned();
         let (tx, mut rx) = mpsc::channel::<BeaconCmd>(1);
-        if is_new_node {
-            info!(parent: &bp.span, "beacon id [{id}]: will run as fresh install -> expect to run DKG.");
-        }
 
         tracker.spawn(async move {
+            let mut seen = SeenPackets::default();
+
             while let Some(cmd) = rx.recv().await {
                 match cmd {
-                    BeaconCmd::Shutdown(callback) => {
-                        if let Err(err) = bp.shutdown(callback) {
-                            error!("failed to shutdown gracefully, {err}")
-                        };
+                    BeaconCmd::IdentityRequest(cb) => cb.reply(bp.identity().try_into()),
+                    BeaconCmd::Sync(cb) => cb.reply(bp.sync()),
+                    BeaconCmd::ChainInfo(cb) => cb.reply(bp.chain_info()),
+                    BeaconCmd::DkgActions(action) => bp.handle_actions(action, &mut seen).await,
+                    BeaconCmd::Shutdown(cb) => {
+                        cb.reply(bp.shutdown());
                         break;
-                    }
-                    BeaconCmd::IdentityRequest(callback) => {
-                        if let Err(err) = bp.identity_request(callback) {
-                            error!("failed to send identity responce, {err}")
-                        }
-                    }
-                    BeaconCmd::Sync(callback) => {
-                        if callback.send(Ok(Arc::clone(&bp.chain_store))).is_err() {
-                            error!("failed to proceed sync request")
-                        }
-                    }
-                    BeaconCmd::ChainInfo(callback) => {
-                        if callback.send(Ok(bp.chain_handler.chain_info())).is_err() {
-                            error!("failed to send chain_info, receiver is dropped")
-                        }
-                    }
-                    BeaconCmd::DkgActions(action) => {
-                        if dkg.send(action).await.is_err() {
-                            error!("{}", crate::dkg::process::ERR_SEND)
-                        }
                     }
                 }
             }
@@ -160,37 +160,60 @@ impl<S: Scheme> BeaconProcess<S> {
         })
     }
 
-    fn shutdown(&self, sender: Callback<(), ()>) -> anyhow::Result<()> {
-        debug!("received shutdown for {}", self.beacon_id);
+    async fn handle_actions(&self, request: Actions, seen: &mut SeenPackets) {
+        match request {
+            Actions::Status(callback) => callback.reply(self.dkg_status()),
+            Actions::Command(cmd, callback) => callback.reply(self.command(cmd).await),
+            Actions::Gossip(packet, callback) => callback.reply(self.packet(packet, seen).await),
+            Actions::Broadcast(bundle, callback) => {
+                callback.reply(todo_request(&bundle.to_string()))
+            }
+        }
+    }
 
-        // TODO: remove 'result' once implemented:
+    fn shutdown(&self) -> Result<(), ShutdownError> {
+        debug!(parent: self.log(), "received shutdown");
+
+        // TODO:
         //  - beacon_id removed from connection pool is_ok
         //  - database closed correctly is_ok
         //  - chain handler stopped correctly is_ok
         //  - any other tasks which might be spawned by tracker.
-        let result = Ok(());
-        sender.send(result).map_err(|_| SendError)?;
 
         Ok(())
     }
 
-    fn identity_request(
-        &self,
-        sender: Callback<IdentityResponse, PointSerDeError>,
-    ) -> anyhow::Result<()> {
-        sender
-            .send(self.keypair.public_identity().try_into())
-            .map_err(|_| SendError)?;
+    fn sync(&self) -> Result<Arc<MemStore>, SyncError> {
+        // TODO: this module is not unfinished
+        Ok(Arc::clone(&self.chain_store))
+    }
 
-        Ok(())
+    fn chain_info(&self) -> Result<ChainInfoPacket, ChainInfoError> {
+        // TODO: this module is unfinished
+        Ok(self.chain_handler.chain_info())
+    }
+
+    /// Returns [`Participant`] which is dkg representation of [`Identity`].
+    pub fn as_participant(&self) -> Result<Participant, ActionsError> {
+        self.identity()
+            .try_into()
+            .map_err(|_| ActionsError::IntoParticipant)
     }
 
     pub fn tracker(&self) -> &TaskTracker {
         &self.tracker
     }
 
-    pub fn id(&self) -> &BeaconID {
-        &self.beacon_id
+    pub fn id(&self) -> &str {
+        self.beacon_id.as_str()
+    }
+
+    pub fn log(&self) -> &Span {
+        &self.log
+    }
+
+    pub fn dkg_store(&self) -> &DkgStore {
+        &self.dkg_store
     }
 
     /// Returns public identity for BeaconID
@@ -219,6 +242,22 @@ impl std::fmt::Display for BeaconID {
 }
 
 #[derive(thiserror::Error, Debug)]
-/// Diagnostic error type.
-#[error("callback receiver has already been dropped")]
-struct SendError;
+#[error("failed to shutdown gracefully")]
+pub struct ShutdownError;
+
+/// Temporary tracker for unfinished logic
+pub fn todo_request<D: std::fmt::Display + ?Sized>(kind: &D) -> Result<(), ActionsError> {
+    tracing::warn!("received TODO request: {kind}");
+
+    Err(ActionsError::TODO)
+}
+
+/// Error type for unfinished sync module
+#[derive(thiserror::Error, Debug)]
+#[error("sync error")]
+pub struct SyncError;
+
+/// Error type for unfinished chain info module
+#[derive(thiserror::Error, Debug)]
+#[error("chain info error")]
+pub struct ChainInfoError;
